fix(sidebar): gate, scope and fully retarget every rename
