+++ Plugins/LibSQLDriverPlugin/LibSQLPlugin.swift
@@ -16,6 +16,7 @@ final class LibSQLPlugin: NSObject, TableProPlugin, DriverPlugin {
     static let databaseTypeId = "libSQL"
 
     static let supportsRenameTable = true
+    static let supportsRenameView = false
     static let additionalDatabaseTypeIds = ["Turso"]
     static let databaseDisplayName = "libSQL / Turso"
     static let iconName = "libsql-icon"
