+++ docs/features/table-operations.mdx
@@ -57,7 +57,7 @@ The statement runs at once instead of joining the drop and truncate queue, so **
 | Dameng | Yes | No | No |
 | DuckDB | Yes | No | No |
 | BigQuery | Yes | No | No |
-| Snowflake | Yes | Yes | Yes |
+| Snowflake | Yes | No | Yes |
 | Trino | Depends on the connector | No | Yes |
 | Teradata | Yes | No | No |
 | MongoDB | Yes | No | No |
